+++ ContentPatcher/Framework/Tokens/Token.cs
@@ -147,6 +147,12 @@ namespace ContentPatcher.Framework.Tokens
         }
 
         /// <inheritdoc />
+        public virtual string NormalizeValue(string value)
+        {
+            return this.Values.NormalizeValue(value);
+        }
+
+        /// <inheritdoc />
         public virtual IEnumerable<string> GetValues(IInputArguments input)
         {
             // get values
@@ -155,7 +161,7 @@ namespace ContentPatcher.Framework.Tokens
             // apply contains
             if (input.ReservedArgs.TryGetValue(InputArguments.ContainsKey, out IInputArgumentValue rawSearch))
             {
-                InvariantHashSet search = new InvariantHashSet(rawSearch.Parsed);
+                InvariantHashSet search = new InvariantHashSet(rawSearch.Parsed.Select(this.NormalizeValue));
                 bool match = search.Any() && values.Any(value => search.Contains(value));
                 values = new[] { match.ToString() };
             }