+++ ContentPatcher/release-notes.md
@@ -11,6 +11,7 @@ When releasing a format change, don't forget to update the smapi.io/json schema!
 -->
 ## Upcoming release
 * Patches can now optionally [update on in-game clock change](docs/author-guide.md#update-rate).
+* Custom mod tokens can now normalize raw values before they're compared with the token.
 
 ## 1.18.6
 Released 21 November 2020.